+++ kroxylicious-integration-test-support/src/test/java/io/kroxylicious/proxy/config/ConfigurationTest.java
@@ -6,6 +6,7 @@
 
 package io.kroxylicious.proxy.config;
 
+import java.time.Duration;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -28,6 +29,9 @@ import io.kroxylicious.proxy.bootstrap.RoundRobinBootstrapSelectionStrategy;
 import io.kroxylicious.proxy.config.tls.TlsClientAuth;
 import io.kroxylicious.proxy.service.HostPort;
 import io.kroxylicious.test.tester.KroxyliciousConfigUtils;
+
+import edu.umd.cs.findbugs.annotations.NonNull;
+import edu.umd.cs.findbugs.annotations.Nullable;
 
 import static io.kroxylicious.test.tester.KroxyliciousConfigUtils.defaultGatewayBuilder;
 import static io.kroxylicious.test.tester.KroxyliciousConfigUtils.defaultPortIdentifiesNodeGatewayBuilder;
@@ -55,65 +59,57 @@ class ConfigurationTest {
 
     @Test
     void shouldRejectVirtualClusterWithNoGateways() {
-        assertThatThrownBy(() -> {
-            MAPPER.readValue(
-                    """
-                              name: cluster
-                              targetCluster:
-                                bootstrapServers: kafka.example:1234
-                            """, VirtualCluster.class);
-        }).isInstanceOf(MismatchedInputException.class)
+        assertThatThrownBy(() -> MAPPER.readValue(
+                """
+                          name: cluster
+                          targetCluster:
+                            bootstrapServers: kafka.example:1234
+                        """, VirtualCluster.class)).isInstanceOf(MismatchedInputException.class)
                 .hasMessageContaining("Missing required creator property 'gateways'");
     }
 
     @Test
     void shouldRejectVirtualClusterWithNullGateways() {
-        assertThatThrownBy(() -> {
-            MAPPER.readValue(
-                    """
-                              name: cluster
-                              targetCluster:
-                                bootstrapServers: kafka.example:1234
-                              gateways: null
-                            """, VirtualCluster.class);
-        }).isInstanceOf(ValueInstantiationException.class)
+        assertThatThrownBy(() -> MAPPER.readValue(
+                """
+                          name: cluster
+                          targetCluster:
+                            bootstrapServers: kafka.example:1234
+                          gateways: null
+                        """, VirtualCluster.class)).isInstanceOf(ValueInstantiationException.class)
                 .hasCauseInstanceOf(IllegalConfigurationException.class)
                 .hasMessageContaining("no gateways configured for virtual cluster 'cluster'");
     }
 
     @Test
     void shouldRejectVirtualClusterNullGatewayValue() {
-        assertThatThrownBy(() -> {
-            MAPPER.readValue(
-                    """
-                              name: cluster
-                              targetCluster:
-                                bootstrapServers: kafka.example:1234
-                              gateways: [null]
-                            """, VirtualCluster.class);
-        }).isInstanceOf(ValueInstantiationException.class)
+        assertThatThrownBy(() -> MAPPER.readValue(
+                """
+                          name: cluster
+                          targetCluster:
+                            bootstrapServers: kafka.example:1234
+                          gateways: [null]
+                        """, VirtualCluster.class)).isInstanceOf(ValueInstantiationException.class)
                 .hasCauseInstanceOf(IllegalConfigurationException.class)
                 .hasMessageContaining("one or more gateways were null for virtual cluster 'cluster'");
     }
 
     @Test
     void shouldRejectSniGatewayWithNoAdvertisedBrokerAddressPattern() {
-        assertThatThrownBy(() -> {
-            MAPPER.readValue(
-                    """
-                              name: cluster
-                              targetCluster:
-                                bootstrapServers: kafka.example:1234
-                              gateways:
-                              - name: default
-                                sniHostIdentifiesNode:
-                                    bootstrapAddress: cluster1:9192
-                                tls:
-                                  key:
-                                    certificateFile: /tmp/cert
-                                    privateKeyFile: /tmp/key
-                            """, VirtualCluster.class);
-        }).isInstanceOf(MismatchedInputException.class)
+        assertThatThrownBy(() -> MAPPER.readValue(
+                """
+                          name: cluster
+                          targetCluster:
+                            bootstrapServers: kafka.example:1234
+                          gateways:
+                          - name: default
+                            sniHostIdentifiesNode:
+                                bootstrapAddress: cluster1:9192
+                            tls:
+                              key:
+                                certificateFile: /tmp/cert
+                                privateKeyFile: /tmp/key
+                        """, VirtualCluster.class)).isInstanceOf(MismatchedInputException.class)
                 .hasMessageContaining("Missing required creator property 'advertisedBrokerAddressPattern'");
     }
 
@@ -502,6 +498,40 @@ class ConfigurationTest {
                                     - name: default
                                       portIdentifiesNode:
                                         bootstrapAddress: example.com:1234
+                                """),
+                argumentSet("UnAuthenticatedIdleTimeout",
+                        new ConfigurationBuilder().addToVirtualClusters(VIRTUAL_CLUSTER).withNewNetwork()
+                                .withNewProxy().withUnauthenticatedIdleTimeout(Duration.ofSeconds(90)).endProxy()
+                                .endNetwork().build(),
+                        """
+                                network:
+                                    proxy:
+                                        unauthenticatedIdleTimeout: 1m30s
+                                virtualClusters:
+                                  - name: demo
+                                    targetCluster:
+                                      bootstrapServers: kafka.example:1234
+                                    gateways:
+                                    - name: default
+                                      portIdentifiesNode:
+                                        bootstrapAddress: example.com:1234
+                                """),
+                argumentSet("AuthenticatedIdleTimeout",
+                        new ConfigurationBuilder().addToVirtualClusters(VIRTUAL_CLUSTER).withNewNetwork()
+                                .withNewProxy().withAuthenticatedIdleTimeout(Duration.ofMinutes(10)).endProxy()
+                                .endNetwork().build(),
+                        """
+                                network:
+                                    proxy:
+                                        authenticatedIdleTimeout: 10m
+                                virtualClusters:
+                                  - name: demo
+                                    targetCluster:
+                                      bootstrapServers: kafka.example:1234
+                                    gateways:
+                                    - name: default
+                                      portIdentifiesNode:
+                                        bootstrapAddress: example.com:1234
                                 """)
 
         );
@@ -524,7 +554,6 @@ class ConfigurationTest {
                 new NamedFilterDefinition("foo", "", ""));
         Optional<Map<String, Object>> development = Optional.empty();
         var virtualCluster = List.of(VIRTUAL_CLUSTER);
-        NetworkDefinition network = null;
         assertThatThrownBy(() -> new Configuration(null,
                 filterDefinitions,
                 null,
@@ -532,7 +561,7 @@ class ConfigurationTest {
                 null,
                 false,
                 development,
-                network))
+                null))
                 .isInstanceOf(IllegalConfigurationException.class)
                 .hasMessage("'filterDefinitions' contains multiple items with the same names: [foo]");
     }
@@ -606,23 +635,9 @@ class ConfigurationTest {
         List<NamedFilterDefinition> filterDefinitions = List.of(
                 new NamedFilterDefinition("foo", "Foo", ""),
                 new NamedFilterDefinition("bar", "Bar", ""));
-        VirtualCluster direct = new VirtualCluster("direct", new TargetCluster("y:9092", Optional.empty()),
-                List.of(new VirtualClusterGateway("mygateway",
-                        new PortIdentifiesNodeIdentificationStrategy(new HostPort("example.com", 3), null, null, null),
-                        null,
-                        Optional.empty())),
-                false,
-                false,
-                List.of("foo")); // filters defined on cluster
 
-        VirtualCluster defaulted = new VirtualCluster("defaulted", new TargetCluster("x:9092", Optional.empty()),
-                List.of(new VirtualClusterGateway("mygateway",
-                        new PortIdentifiesNodeIdentificationStrategy(new HostPort("example.com", 3), null, null, null),
-                        null,
-                        Optional.empty())),
-                false,
-                false,
-                null); // filters not defined => should default to the top level
+        VirtualCluster direct = buildVirtualCluster("direct", "y:9092", List.of("foo")); // filters defined on cluster
+        VirtualCluster defaulted = buildVirtualCluster("defaulted", "x:9092", null); // filters not defined => should default to the top level
 
         Configuration configuration = new Configuration(
                 null,
@@ -638,10 +653,24 @@ class ConfigurationTest {
         var model = configuration.virtualClusterModel();
 
         // Then
-        var directModel = model.stream().filter(x -> x.getClusterName().equals("direct")).findFirst().get();
-        var defaultModel = model.stream().filter(x -> x.getClusterName().equals("defaulted")).findFirst().get();
-        assertThat(directModel.getFilters()).singleElement().extracting(NamedFilterDefinition::type).isEqualTo("Foo");
-        assertThat(defaultModel.getFilters()).singleElement().extracting(NamedFilterDefinition::type).isEqualTo("Bar");
+        assertThat(model.stream().filter(x -> x.getClusterName().equals("direct")).findFirst())
+                .isPresent()
+                .hasValueSatisfying(vcm -> assertThat(vcm.getFilters()).singleElement().extracting(NamedFilterDefinition::type).isEqualTo("Foo"));
+        assertThat(model.stream().filter(x -> x.getClusterName().equals("defaulted")).findFirst())
+                .isPresent()
+                .hasValueSatisfying(vcm -> assertThat(vcm.getFilters()).singleElement().extracting(NamedFilterDefinition::type).isEqualTo("Bar"));
+    }
+
+    @NonNull
+    private static VirtualCluster buildVirtualCluster(String virtualClusterName, String targetBootstrap, @Nullable List<String> filterNames) {
+        return new VirtualCluster(virtualClusterName, new TargetCluster(targetBootstrap, Optional.empty()),
+                List.of(new VirtualClusterGateway("mygateway",
+                        new PortIdentifiesNodeIdentificationStrategy(new HostPort("example.com", 3), null, null, null),
+                        null,
+                        Optional.empty())),
+                false,
+                false,
+                filterNames);
     }
 
 }